Allow user to provide path to spm.m
The user should provide the path to the SPM folder.
In case they provide the path to spm.m (e.g., by drag-and-dropping the
file onto the command line), automatically extract folder and handle the
situation from there.
Added extra logic to check if file exists, if file is valid, same for
folder.